% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuBA.R
\name{BiclusterGenesGraph}
\alias{BiclusterGenesGraph}
\title{Bicluster Genes Graph Function}
\usage{
BiclusterGenesGraph(BiclusterGenes, GenePairs, GeneNames, BiclusterNos)
}
\arguments{
\item{BiclusterGenes}{A character variable. Specifies the name of the .csv file containing the lists of genes in the biclusters generated by the \link[TuBA]{Biclustering} function.}

\item{GenePairs}{A character variable. Specifies the name of the genes-pairs .csv file generated by the \link[TuBA]{GenePairs} function.}

\item{GeneNames}{A character variable. Specifies the name of the .csv file containing the gene names/IDs generated by the \link[TuBA]{GenePairs} function.}

\item{BiclusterNos}{A numeric vector. Specifies the serial numbers of the biclusters for which the bicluster graphs/networks are desired.}
}
\value{
Generates .pdf files containing the graphs/networks of genes corresponding to the biclusters specified by the user.
}
\description{
This function can be used to make graphs/networks containing the genes in the biclusters found by the \link[TuBA]{Biclustering} function.
}
\examples{
\dontrun{
BiclusterGenesGraph(BiclusterGenes = "RPGenes_H0.05_JcdInd0.2_GenesInBiclusters.csv",GenePairs = "RPGenes_H0.05_JcdInd0.2_GenePairs.csv",GeneNames = "RPGenes_H0.05_JcdInd0.2_GeneNames.csv",BiclusterNos = c(1,2))
}
}
