% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuBA.R
\name{Biclustering}
\alias{Biclustering}
\title{Biclustering Function}
\usage{
Biclustering(
  VariablePairs,
  BinaryMatrix,
  MinGenes = NULL,
  MinSamples = NULL,
  SampleEnrichment = NULL
)
}
\arguments{
\item{VariablePairs}{A character variable. Specifies the name of the gene-pairs .csv file generated by the \link[TuBA]{GenePairs} function.}

\item{BinaryMatrix}{A character variable. Specifies the name of the genes-samples binary matrix .csv file generated by the \link[TuBA]{GenePairs} function.}

\item{MinGenes}{A numeric (integer) variable. Specifies the minimum number of genes that the biclusters must contain.}

\item{MinSamples}{A numeric (integer) variable. Specifies the minimum number of samples that the biclusters must contain.}

\item{SampleEnrichment}{A numeric variable - should be greater than 0 and less than or equal to 1. Specifies to what extent the samples should be enriched in the bicluster. Smaller values indicate higher enrichment. Default is 1.}
}
\value{
A data frame containing the sets of genes in the biclusters along with the information about the number of samples in each bicluster. In addition, it generates 3 .csv files - one contains the list of genes in each bicluster along with information about the total number of samples in the bicluster and how many of them are contributed by each gene within the bicluster; the other contains the bicluster-samples binary matrix (biclusters along rows, samples along columns) in which the presence of a sample within a bicluster is indicated by a 1; the third file contains the genes-biclusters-samples matrix (genes along rows, samples along columns) which contains more detailed information about which sample is present for which gene within a given bicluster. The first column in this genes-biclusters-samples file contains the names of genes in the bicluster, and the second column contains the serial numbers of the biclusters which contains these genes; the rest of the file contains the binary matrix.
}
\description{
This function finds the biclusters using the graph (gene-pairs file) and the genes-samples binary matrix generated by the \link[TuBA]{GenePairs} function.
}
\examples{
\dontrun{
Biclustering(VariablePairs = "RPGenes_H0.05_JcdInd0.2_GenePairs.csv",BinaryMatrix = "RPGenes_H0.05_JcdInd0.2_GenesSamples_BinaryMatrix.csv")
}
}
