% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuBA.R
\name{DataCleaning}
\alias{DataCleaning}
\title{DataCleaning}
\usage{
DataCleaning(File)
}
\arguments{
\item{File}{A character variable. Specifies the name of the .csv or .txt file that contains the gene expression data. Make sure it is in the genes along rows and samples along columns format.}
}
\value{
A data frame containing the gene expression data after basic cleaning. In addition, it generates two .csv files in the working directory - one contains the names of the genes in the cleaned gene expression data set, and the other contains the entire gene expression data after cleaning.
}
\description{
This function performs basic cleaning of gene expression data sets - removes genes with zero counts across all samples as well as genes with NAs in some samples.
}
\examples{
\dontrun{
DataCleaning(File = "RPGenes.csv")
}
}
