% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuBA.R
\name{DataPrep}
\alias{DataPrep}
\title{DataPrep}
\usage{
DataPrep(X, normalize = T)
}
\arguments{
\item{X}{A character variable or a matrix. Specifies the name of the file that contains the raw counts data (should be in .csv, .txt, or .tsv format). Make sure it is in the genes along rows and samples along columns format. Alternatively, you can provide a matrix with raw counts as input. The matrix must contain the Gene IDs as rownames and Sample IDs as colnames.}

\item{normalize}{A logical variable. Specifies whether the raw counts should be normalized using the DESeq normalization method (default is T).}
}
\value{
A matrix containing the gene expression data after basic cleaning and normalization (if normalize = T). In addition, it generates a .csv file in the working directory that contains the cleaned and normalized (if normalize = T) gene expression data set.
}
\description{
This function performs basic cleaning of raw counts data sets - removes genes with zero counts across all samples, as well as genes with NAs in some samples. Additionally, it offers the option (default is T) to perform between-sample normalization using the DESeq approach.
}
\examples{
\dontrun{
DataPrep(X = "RPGenes.csv",normalize = F)
}
}
